refactor: 메인과 테스트 패키지 레벨 동일화
