+++ src/main/java/blackjack/domain/BettingAmount.java
@@ -1,4 +1,4 @@
-package blackjack.domain.deck;
+package blackjack.domain;
 
 public class BettingAmount {
