+++ src/test/java/blackjack/domain/BettingAmountTest.java
@@ -2,7 +2,6 @@ package blackjack.domain;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import blackjack.domain.deck.BettingAmount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
